brainflow/requirements.txt
pyyaml>=6.0.1
pydantic>=2.6
rich>=13.7
pillow>=10.4
